Agregue DTOs y corregi detalles de las entidades pt 3

+++ src/main/java/utec/proyectofinal/Proyecto/Final/UTEC/business/entities/Analisis.java
@@ -15,7 +15,7 @@ public abstract class Analisis {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer analisisID;
+    private Long analisisID;
 
     @ManyToOne
     @JoinColumn(name = "loteID")
